requirements.txt
openai>=1.0.0
inquirer>=3.1.3
pyyaml>=6.0
